backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
requests==2.33.1
openai-whisper==20231117
sentence-transformers==2.7.0
spacy>=3.7.4
scikit-learn==1.4.2
numpy==1.26.4
moviepy==1.0.3
python-pptx==0.6.23
PyMuPDF==1.24.3
pdfplumber==0.11.0
nltk==3.8.1
aiofiles==23.2.1
python-dotenv==1.0.1
google-genai>=0.3.0
pydantic>=2.0.0
